+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - Factorize the code creating roles & permissions for organizations & courses.
+- Add icons to the course detail page via a constrained placeholder,
 - Add an icon to the category detail page via a constrained placeholder.
 
 ### Fixed